S3 terminator updates
Move S3Bucket class to a new storage_services.py (we previously moved the
policies) and add a new class to handle emptying of a new, persistent,
encrypted bucket to be used for SSM connection plugin tests
